+++ v2/ui/src/CommandBuilder.h
@@ -108,7 +108,7 @@ public:
         PushFloat(clip_width);
         PushFloat(clip_height);
         const std::uint32_t bounds_flags =
-            (interactive ? ED_BOUNDS_FLAG_INTERACTIVE : 0U) |
+            (interactive ? static_cast<std::uint32_t>(ED_BOUNDS_FLAG_INTERACTIVE) : 0U) |
             ((clip_mode << ED_BOUNDS_CLIP_MODE_SHIFT) & ED_BOUNDS_CLIP_MODE_MASK);
         words_.push_back(bounds_flags);
     }
